HIGH FIX PROPAGATED: partial changes in the PU's
