+++ src/main/java/com/network/franchise/application/config/UseCaseConfig.java
@@ -1,9 +1,6 @@
 package com.network.franchise.application.config;
 
 import com.network.franchise.domain.api.AppPersistenceAdapterPort;
-import com.network.franchise.infrastructure.inbound.mapper.BranchesMapper;
-import com.network.franchise.infrastructure.inbound.mapper.FranchiseMapper;
-import com.network.franchise.infrastructure.inbound.mapper.ProductsMapper;
 import com.network.franchise.domain.usecase.command.*;
 import com.network.franchise.domain.usecase.queries.GetTopProductsPerBranchUseCase;
 import org.springframework.context.annotation.Bean;
@@ -20,9 +17,8 @@ public class UseCaseConfig {
     }
 
     @Bean
-    public AddBranchUseCase addBranchUseCase(AppPersistenceAdapterPort appPersistenceAdapterPort,
-                                             BranchesMapper mapper) {
-        return new AddBranchUseCase(appPersistenceAdapterPort, mapper);
+    public AddBranchUseCase addBranchUseCase(AppPersistenceAdapterPort appPersistenceAdapterPort) {
+        return new AddBranchUseCase(appPersistenceAdapterPort);
     }
 
     @Bean
@@ -46,8 +42,7 @@ public class UseCaseConfig {
     }
 
     @Bean
-    public UpdateFranchiseNameUseCase updateFranchiseNameUseCase(AppPersistenceAdapterPort appPersistenceAdapterPort,
-                                                                 FranchiseMapper mapper) {
-        return new UpdateFranchiseNameUseCase(appPersistenceAdapterPort, mapper);
+    public UpdateFranchiseNameUseCase updateFranchiseNameUseCase(AppPersistenceAdapterPort appPersistenceAdapterPort) {
+        return new UpdateFranchiseNameUseCase(appPersistenceAdapterPort);
     }
 }
