+++ src/main/java/com/network/franchise/domain/usecase/command/UpdateFranchiseNameUseCase.java
@@ -4,7 +4,7 @@ import com.network.franchise.domain.api.AppPersistenceAdapterPort;
 import com.network.franchise.domain.common.enums.TechnicalMessage;
 import com.network.franchise.domain.common.exceptions.BusinessException;
 import com.network.franchise.domain.common.exceptions.DuplicateException;
-import com.network.franchise.infrastructure.inbound.mapper.FranchiseMapper;
+import com.network.franchise.domain.mapper.FranchiseDtoMapper;
 import com.network.franchise.domain.model.Franchise;
 import com.network.franchise.domain.spi.UpdateFranchiseNameServicePort;
 import reactor.core.publisher.Mono;
@@ -12,11 +12,9 @@ import reactor.core.publisher.Mono;
 public class UpdateFranchiseNameUseCase implements UpdateFranchiseNameServicePort {
 
     private final AppPersistenceAdapterPort appPersistenceAdapterPort;
-    private final FranchiseMapper mapper;
 
-    public UpdateFranchiseNameUseCase(AppPersistenceAdapterPort appPersistenceAdapterPort, FranchiseMapper mapper) {
+    public UpdateFranchiseNameUseCase(AppPersistenceAdapterPort appPersistenceAdapterPort) {
         this.appPersistenceAdapterPort = appPersistenceAdapterPort;
-        this.mapper = mapper;
     }
 
     @Override
@@ -27,7 +25,7 @@ public class UpdateFranchiseNameUseCase implements UpdateFranchiseNameServicePor
         return appPersistenceAdapterPort.existsFranchiseByName(request.getName())
                 .flatMap(exists -> {
                     if (exists) return Mono.error(new DuplicateException(TechnicalMessage.ALREADY_EXISTS));
-                    return appPersistenceAdapterPort.updateFranchiseName(mapper.toEntityFromDomainFranchise(request));
+                    return appPersistenceAdapterPort.updateFranchiseName(FranchiseDtoMapper.INSTANCE.toEntityFromDomainFranchise(request));
                 })
                 .switchIfEmpty(Mono.error(new BusinessException(TechnicalMessage.BAD_REQUEST)));
     }
