+++ src/main/java/com/network/franchise/domain/usecase/command/UpdateStockUseCase.java
@@ -6,7 +6,6 @@ import com.network.franchise.domain.common.exceptions.BusinessException;
 import com.network.franchise.domain.common.exceptions.NotFoundException;
 import com.network.franchise.domain.dto.response.CreateProductResponseDto;
 import com.network.franchise.domain.mapper.ProductDtoMapper;
-import com.network.franchise.infrastructure.inbound.mapper.ProductsMapper;
 import com.network.franchise.domain.model.Product;
 import com.network.franchise.domain.spi.UpdateStockServicePort;
 import reactor.core.publisher.Mono;
@@ -14,7 +13,6 @@ import reactor.core.publisher.Mono;
 public class UpdateStockUseCase implements UpdateStockServicePort {
 
     private final AppPersistenceAdapterPort appPersistenceAdapterPort;
-//    private final ProductsMapper mapper;
 
     public UpdateStockUseCase(AppPersistenceAdapterPort appPersistenceAdapterPort) {
         this.appPersistenceAdapterPort = appPersistenceAdapterPort;
