+++ src/main/java/com/network/franchise/infrastructure/inbound/mapper/BranchesMapper.java
@@ -1,8 +1,6 @@
 package com.network.franchise.infrastructure.inbound.mapper;
 
-import com.network.franchise.domain.dto.response.CreateBranchResponseDto;
 import com.network.franchise.domain.model.Branch;
-import com.network.franchise.infrastructure.adapters.persistence.entity.BranchEntity;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -13,23 +11,6 @@ public class BranchesMapper {
         return Branch.builder()
                 .franchiseId(franchiseId)
                 .name(branch.getName())
-                .build();
-    }
-
-    public BranchEntity toEntityFromDomainBranch(Branch domain) {
-        if (domain == null) return null;
-        return BranchEntity.builder()
-                .name(domain.getName())
-                .franchiseId(domain.getFranchiseId())
-                .build();
-    }
-
-    public CreateBranchResponseDto toDomainFromBranchEntity(BranchEntity entity) {
-        if (entity == null) return null;
-        return CreateBranchResponseDto.builder()
-                .id(entity.getId())
-                .name(entity.getName())
-                .franchiseId(entity.getFranchiseId())
                 .build();
     }
 }
